Gradle build
-Split library and cli project
-Add depenency definition in gradle configuration
-Delete dependency libraries
-Drop netbeans files
-Drop Ant files
