+++ build.gradle
@@ -1,0 +1,33 @@
+// common settings
+subprojects {
+  apply plugin: 'java'
+  apply plugin: "jacoco"
+
+  sourceCompatibility = '1.7'
+  targetCompatibility = '1.7'
+  [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
+  [compileJava, compileTestJava]*.options*.bootClasspath = "$JDK7_HOME/jre/lib/rt.jar"
+
+  repositories {
+    mavenCentral()
+  }
+
+  dependencies {
+    testCompile group: 'junit', name: 'junit', version: '4.11'
+  }
+}
+
+project(':dictzip-lib') {}
+
+project(':dictzip-cli') {
+  apply plugin: 'application'
+  mainClassName = 'org.dict.zip.cli.Main'
+  applicationName = 'DictZip'
+
+  dependencies {
+    compile project(':dictzip-lib'),
+       'commons-codec:commons-codec:1.10',
+       'gnu.getopt:java-getopt:1.0.13'
+    testCompile 'commons-io:commons-io:2.4'
+  }
+}
