--- build.xml
@@ -1,155 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- You may freely edit this file. See commented blocks below for -->
-<!-- some examples of how to customize the build. -->
-<!-- (If you delete it and reopen the project it will be recreated.) -->
-<!-- By default, only the Clean and Build commands use this build script. -->
-<!-- Commands such as Run, Debug, and Test only use this build script if -->
-<!-- the Compile on Save feature is turned off for the project. -->
-<!-- You can turn off the Compile on Save (or Deploy on Save) setting -->
-<!-- in the project's Project Properties dialog box.-->
-<project name="dictzip" default="default" basedir=".">
-    <description>Builds, tests, and runs the project dictzip.</description>
-    <taskdef resource="com/puppycrawl/tools/checkstyle/ant/checkstyle-ant-task.properties"
-      classpath="lib/checkstyle-6.13-all.jar"/>
-    <import file="nbproject/build-impl.xml"/>
-    <!--
-
-    There exist several targets which are by default empty and which can be 
-    used for execution of your tasks. These targets are usually executed 
-    before and after some main targets. They are: 
-
-      -pre-init:                 called before initialization of project properties
-      -post-init:                called after initialization of project properties
-      -pre-compile:              called before javac compilation
-      -post-compile:             called after javac compilation
-      -pre-compile-single:       called before javac compilation of single file
-      -post-compile-single:      called after javac compilation of single file
-      -pre-compile-test:         called before javac compilation of JUnit tests
-      -post-compile-test:        called after javac compilation of JUnit tests
-      -pre-compile-test-single:  called before javac compilation of single JUnit test
-      -post-compile-test-single: called after javac compilation of single JUunit test
-      -pre-jar:                  called before JAR building
-      -post-jar:                 called after JAR building
-      -post-clean:               called after cleaning build products
-
-    (Targets beginning with '-' are not intended to be called on their own.)
-
-    Example of inserting an obfuscator after compilation could look like this:
-
-        <target name="-post-compile">
-            <obfuscate>
-                <fileset dir="${build.classes.dir}"/>
-            </obfuscate>
-        </target>
-
-    For list of available properties check the imported 
-    nbproject/build-impl.xml file. 
-
-
-    Another way to customize the build is by overriding existing main targets.
-    The targets of interest are: 
-
-      -init-macrodef-javac:     defines macro for javac compilation
-      -init-macrodef-junit:     defines macro for junit execution
-      -init-macrodef-debug:     defines macro for class debugging
-      -init-macrodef-java:      defines macro for class execution
-      -do-jar:                  JAR building
-      run:                      execution of project 
-      -javadoc-build:           Javadoc generation
-      test-report:              JUnit report generation
-
-    An example of overriding the target for project execution could look like this:
-
-        <target name="run" depends="dictzip-impl.jar">
-            <exec dir="bin" executable="launcher.exe">
-                <arg file="${dist.jar}"/>
-            </exec>
-        </target>
-
-    Notice that the overridden target depends on the jar target and not only on 
-    the compile target as the regular run target does. Again, for a list of available 
-    properties which you can use, check the target you are overriding in the
-    nbproject/build-impl.xml file. 
-
-    -->
-    <!-- Set the repo.revision property from the first valid source in this order:
-     1. src/**/Version.properties
-     2. This git SHA1
-    -->
-  <target name="get-revision" depends="get-source-revision,get-git-revision">
-    <!-- Fallback in case value was not set from source or git -->
-    <property name="repo.revision" value="unknown"/>
-    <echo message="Detected revision: ${repo.revision}"/>
-  </target>
-
-  <target name="get-source-revision">
-    <condition property="repo.revision" value="${revision}">
-      <and>
-        <isset property="revision"/>
-        <not>
-          <equals arg1="${revision}" arg2="dev"/>
-        </not>
-      </and>
-    </condition>
-  </target>
-
-  <available file=".git" type="dir" property="git.available"/>
-  <target name="get-git-revision" if="git.available">
-    <exec executable="git" outputproperty="git.sha1" failifexecutionfails="false">
-      <arg value="rev-parse"/>
-      <arg value="--short"/>
-      <arg value="HEAD"/>
-    </exec>
-    <condition property="repo.revision" value="${git.sha1}">
-      <and>
-        <isset property="git.sha1"/>
-        <length string="${git.sha1}" trim="yes" length="0" when="greater"/>
-      </and>
-    </condition>
-  </target>
-
-  <!-- Builds a source distribution -->
-  <target name="src" depends="init,get-revision" description="Build source distribution">
-    <mkdir dir="dist2"/>
-    <!-- Create temporary Version.properties with current revision -->
-    <tempfile property="temp.version.props" destdir="dist2" deleteonexit="true"/>
-    <copy file="src/main/resources/org/dict/zip/Version.properties" tofile="${temp.version.props}"/>
-    <propertyfile file="${temp.version.props}">
-      <entry key="revision" value="${repo.revision}"/>
-    </propertyfile>
-    <!-- Build zip -->
-    <zip destfile="dist2/dictzip_${version}_${revision}_Source.zip">
-      <fileset dir="."
-          includes="doc/** lib/** nbproject/** src/** build.xml"
-          excludes="**/nbproject/private/** **/build/** **/dist/** src/**/Version.properties" />
-      <zipfileset file="${temp.version.props}" fullpath="src/main/resources/org/dict/zip/Version.properties"/>
-    </zip>
-  </target>
-
-
-  <target name="-post-compile-test" depends="checkstyle"/>
-
-  <target name="-pre-jar">
-    <jar compress="${jar.compress}" index="${jar.index}" jarfile="dist/dictzip-lib.jar">
-      <fileset dir="${build.classes.dir}" excludes="**/cli/*,**/cli/"/>
-      <manifest>
-        <attribute name="Built-By" value="${user.name}"/>
-    </manifest>
-    </jar>
-  </target>
-
-  <target name="checkstyle"
-        description="Generates a report of code convention violations.">
-    <checkstyle config="doc/codingstyle_checks.xml">
-      <fileset dir="src">
-        <patternset>
-          <include name="**/*.java"/>
-          <exclude name="**/*Test*"/>
-        </patternset>
-      </fileset>
-      <formatter type="plain"/>
-      <formatter type="xml" toFile="build/checkstyle_errors.xml"/>
-    </checkstyle>
-  </target>
-
-</project>
