+++ dictzip-cli/doc/dictzip.1.in
