+++ dictzip-cli/src/main/java/org/dict/zip/cli/DictData.java
