+++ dictzip-cli/src/test/resources/test.dict.dz
