+++ dictzip-cli/src/test/resources/test2.dict
