+++ dictzip-lib/src/main/java/org/dict/zip/DictZipHeader.java
