+++ dictzip-lib/src/main/java/org/dict/zip/DictZipInputStream.java
