+++ dictzip-lib/src/main/java/org/dict/zip/DictZipOutputStream.java
