+++ dictzip-lib/src/main/java/org/dict/zip/Version.properties
