--- nbproject/project.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://www.netbeans.org/ns/project/1">
-    <type>org.netbeans.modules.java.j2seproject</type>
-    <configuration>
-        <data xmlns="http://www.netbeans.org/ns/j2se-project/3">
-            <name>dictzip</name>
-            <source-roots>
-                <root id="src.dir"/>
-                <root id="src.resources.dir" name="リソース"/>
-            </source-roots>
-            <test-roots>
-                <root id="test.src.dir"/>
-                <root id="test.resources.dir" name="テスト・リソース"/>
-            </test-roots>
-        </data>
-    </configuration>
-</project>
